requirements.txt
streamlit
google-generativeai
pillow
python-dotenv
tenacity
requests
gTTS
moviepy
edge-tts
